requirements.txt
Flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
flask_cors
googleapiclient.discovery